feature to provide data for alignmentFirstPublishedDate
