+++ src/indexClinicalData/types.ts
@@ -8,6 +8,11 @@ export type DonorMolecularDataProcessingStatus =
   | "PROCESSING"
   | "REGISTERED";
 
+export enum FirstPublishedDateFields {
+  RAW_READS_FIRST_PUBLISHED_DATE = "rawReadsFirstPublishedDate",
+  ALIGNMENT_FIRST_PUBLISHED_DATE = "alignmentFirstPublishedDate",
+}
+
 export type RdpcDonorInfo = {
   publishedNormalAnalysis: number;
   publishedTumourAnalysis: number;
@@ -16,6 +21,7 @@ export type RdpcDonorInfo = {
   alignmentsCompleted: number;
   alignmentsRunning: number;
   alignmentsFailed: number;
+  alignmentFirstPublishedDate?: Date;
 
   sangerVcsCompleted: number;
   sangerVcsRunning: number;
